+++ test/storeSpec.js
@@ -13,7 +13,8 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 
 describe('store', () => {
-  it('assembles a store', () => {
+  it('assembles a store', function () {
+    this.expected  = 5;
     const initial  = {};
     const fooStore = store({
       'EVENT_NAME': {
@@ -26,7 +27,9 @@ describe('store', () => {
   });
 
   describe('state handling', () => {
-    it('uses generic initializer which will be used lazily', () => {
+    it('uses generic initializer which will be used lazily', function () {
+      this.expected = 15;
+
       const initializer = sinon.spy();
       const newStore    = store({}, initializer);
 
@@ -40,7 +43,9 @@ describe('store', () => {
     });
 
     describe('property path evaluation of state objects', () => {
-      it('fetches state by evaluating a property path', () => {
+      it('fetches state by evaluating a property path', function () {
+        this.expected = 3;
+
         const newStore = store({}, {
           foo: [
             { bar: 'any-value' }
@@ -50,7 +55,9 @@ describe('store', () => {
         expect(newStore.getStateValue('foo.0.bar')).to.equal('any-value');
       });
 
-      it('fetches state from an array in a property path semantically correct as described in #33', () => {
+      it('fetches state from an array in a property path semantically correct as described in #33', function () {
+        this.expected = 3;
+
         const newStore = store({}, {
           foo: [
             { bar: { blah: 'any-value' } }
@@ -60,7 +67,9 @@ describe('store', () => {
         expect(newStore.getStateValue('foo[0].bar.blah')).to.equal('any-value');
       });
 
-      it ('fetches state from an array where the array is the first level of the state', () => {
+      it ('fetches state from an array where the array is the first level of the state', function () {
+        this.expected = 2;
+
         const newStore = store({}, [
           [
             { bar: { blah: 'any-value' } }
@@ -70,13 +79,16 @@ describe('store', () => {
         expect(newStore.getStateValue('[0][0].bar.blah')).to.equal('any-value');
       });
 
-      it('handles simple array expression', () => {
+      it('handles simple array expression', function () {
+        this.expected = 2;
+
         const newStore = store({}, [0]);
 
         expect(newStore.getStateValue('[0]')).to.equal(0);
       });
 
-      it('handles array expression at start and end', () => {
+      it('handles array expression at start and end', function () {
+        this.expected = 2;
         const newStore = store({}, [{
           bar: ['foo']
         }]);
@@ -85,13 +97,16 @@ describe('store', () => {
       });
     });
 
-    it('returns default value in case of invalid property path', () => {
+    it('returns default value in case of invalid property path', function () {
+      this.expected = 2;
+
       const newStore = store({}, {});
 
       expect(newStore.getStateValue('foo.property', 'default')).to.equal('default');
     });
 
-    it('throws error if proeprty path evaluation should be computed on a scalar', () => {
+    it('throws error if proeprty path evaluation should be computed on a scalar', function () {
+      this.expected = 2;
       expect(() => store({}, 'BLAH').getStateValue('foo.property')).to.throw('To evaluate a property path, the value must be an object or an array!');
     });
   });